+++ src/test/application/diagnostics/checks/pythonInterpreter.unit.test.ts
@@ -29,7 +29,12 @@ import { CommandsWithoutArgs } from '../../../../client/common/application/comma
 import { IPlatformService } from '../../../../client/common/platform/types';
 import { IConfigurationService, IDisposableRegistry, IPythonSettings } from '../../../../client/common/types';
 import { noop } from '../../../../client/common/utils/misc';
-import { IInterpreterHelper, IInterpreterService, InterpreterType } from '../../../../client/interpreter/contracts';
+import {
+    IInterpreterHelper,
+    IInterpreterService,
+    InterpreterType,
+    PythonInterpreter
+} from '../../../../client/interpreter/contracts';
 import { IServiceContainer } from '../../../../client/ioc/types';
 
 suite('Application Diagnostics - Checks Python Interpreter', () => {
@@ -129,7 +134,7 @@ suite('Application Diagnostics - Checks Python Interpreter', () => {
             expect(diagnostics).to.be.deep.equal([]);
             settings.verifyAll();
         });
-        test('Should return diagnostics if there are no interpreters', async () => {
+        test('Should return diagnostics if there are no interpreters after double-checking', async () => {
             settings
                 .setup((s) => s.disableInstallationChecks)
                 .returns(() => false)
@@ -138,12 +143,44 @@ suite('Application Diagnostics - Checks Python Interpreter', () => {
                 .setup((i) => i.hasInterpreters)
                 .returns(() => Promise.resolve(false))
                 .verifiable(typemoq.Times.once());
+            interpreterService
+                .setup((i) => i.getInterpreters(undefined))
+                .returns(() => Promise.resolve([]))
+                .verifiable(typemoq.Times.once());
 
             const diagnostics = await diagnosticService.diagnose(undefined);
             expect(diagnostics).to.be.deep.equal(
                 [new InvalidPythonInterpreterDiagnostic(DiagnosticCodes.NoPythonInterpretersDiagnostic, undefined)],
                 'not the same'
             );
+            settings.verifyAll();
+            interpreterService.verifyAll();
+        });
+        test('Should return empty diagnostics if there are interpreters after double-checking', async () => {
+            const interpreter: PythonInterpreter = { type: InterpreterType.Unknown } as any;
+
+            settings
+                .setup((s) => s.disableInstallationChecks)
+                .returns(() => false)
+                .verifiable(typemoq.Times.once());
+            interpreterService
+                .setup((i) => i.hasInterpreters)
+                .returns(() => Promise.resolve(false))
+                .verifiable(typemoq.Times.once());
+            interpreterService
+                .setup((i) => i.getInterpreters(undefined))
+                .returns(() => Promise.resolve([interpreter]))
+                .verifiable(typemoq.Times.once());
+            interpreterService
+                .setup((i) => i.getActiveInterpreter(typemoq.It.isAny()))
+                .returns(() => {
+                    return Promise.resolve(interpreter);
+                })
+                .verifiable(typemoq.Times.once());
+
+            const diagnostics = await diagnosticService.diagnose(undefined);
+
+            expect(diagnostics).to.be.deep.equal([], 'not the same');
             settings.verifyAll();
             interpreterService.verifyAll();
         });